Remove EDH donations api methods & tests
